+++ testData/org/elixir_lang/model/psi/function/beam_capture_goto.ex
@@ -1,0 +1,5 @@
+defmodule BeamCaptureGoto do
+  def run do
+    &:queue.ne<caret>w/0
+  end
+end