+++ testData/org/elixir_lang/model/psi/function/goto_declaration_qualified_call.ex
@@ -1,0 +1,5 @@
+defmodule GotoDeclarationQualifiedCall do
+  def run(struct, params) do
+    GotoDeclarationCapture.Referenced.change<caret>set(struct, params)
+  end
+end