+++ testData/org/elixir_lang/model/psi/function/goto_declaration_qualified_capture.ex
@@ -1,0 +1,5 @@
+defmodule GotoDeclarationQualifiedCapture do
+  def run do
+    &GotoDeclarationCapture.Referenced.change<caret>set/2
+  end
+end