+++ testData/org/elixir_lang/model/psi/function/usages_capture.ex
@@ -1,0 +1,9 @@
+defmodule UsagesCapture do
+  def per<caret>form(data) do
+    data
+  end
+end
+
+defmodule UsagesCaptureCaller do
+  def run, do: &UsagesCapture.perform/1
+end